detach Model part from Experiments.jl
