+++ examples/Experiments.jl
@@ -1,5 +1,8 @@
 module Experiments
 
+include("Models.jl")
+using .Models
+
 using MLJ
 using MLJ: partition, accuracy
 using MLJFlux
@@ -16,22 +19,15 @@ using BSON
 using Dates
 
 export ExperimentConfig, run_experiment, load_mnist_for_mlj, flatten_images
-export make_mnist_cnn, extract_flux_model
-export make_mnist_forest, make_mnist_tree
-export blackbox_predict
+export make_mnist_cnn, make_mnist_forest, make_mnist_tree
+export blackbox_predict, extract_flux_model
 export save_experiment_result, load_experiment_result, get_or_train_cnn
-
-const DecisionTreeClassifier = @load DecisionTreeClassifier pkg = DecisionTree
-const RandomForestClassifier = @load RandomForestClassifier pkg = DecisionTree
-
-const ImageClassifier = MLJFlux.ImageClassifier
 
 const MODELS_DIR = joinpath(@__DIR__, ".", "models")
 
 # =========================
-# Config
+# Config & Data
 # =========================
-
 struct ExperimentConfig
     name::String
     train_fraction::Float64
@@ -104,6 +100,7 @@ function train_test_split(n::Integer; fraction_train::Float64, rng::Int)
     train, test = partition(1:n, fraction_train, shuffle = true, rng = rng)
     return train, test
 end
+
 """
     run_experiment(model, X, y; config=DEFAULT_CONFIG)
 
@@ -128,15 +125,12 @@ function run_experiment(model, X, y; config::ExperimentConfig = DEFAULT_CONFIG)
     # For DataFrame inputs we must use two-dimensional indexing
     Xtrain = X isa DataFrame ? X[train, :] : X[train]
     Xtest = X isa DataFrame ? X[test, :] : X[test]
-
     mach = machine(model, Xtrain, y[train])
     fit!(mach, verbosity = 1)
-
     ŷ_test = predict(mach, Xtest)
     acc = accuracy(mode.(ŷ_test), y[test])
 
     report = (accuracy = acc,)
-
     return (
         mach = mach,
         train_idx = train,
@@ -146,137 +140,6 @@ function run_experiment(model, X, y; config::ExperimentConfig = DEFAULT_CONFIG)
         config = config,
         report = report,
     )
-end
-
-# =========================
-# Traditional ML Models (Black-box)
-# =========================
-
-"""
-    make_mnist_forest(; rng=42, n_trees=100, max_depth=-1)
-
-Construct a `RandomForestClassifier` suitable as a black-box
-baseline on flattened MNIST features.
-
-- `n_trees`: number of trees in the ensemble.
-- `max_depth`: maximum depth of each tree (-1 means unlimited).
-"""
-function make_mnist_forest(; rng::Int = 42, n_trees::Int = 100, max_depth::Int = -1)
-    model = RandomForestClassifier(n_trees = n_trees, max_depth = max_depth, rng = rng)
-    return model
-end
-
-"""
-    make_mnist_tree(; rng=42, max_depth=5)
-
-Construct a single `DecisionTreeClassifier` for MNIST features.
-Useful as a simpler black-box baseline.
-"""
-function make_mnist_tree(; rng::Int = 42, max_depth::Int = 5)
-    model = DecisionTreeClassifier(max_depth = max_depth, rng = rng)
-    return model
-end
-
-"""
-    blackbox_predict(mach, X)
-
-Pure prediction API for black-box attacks. Given an MLJ machine
-and new features `X`, return probabilistic predictions.
-
-This intentionally hides all training details and gradients.
-"""
-function blackbox_predict(mach, X)
-    # For tree/forest models, `predict` already returns probabilistic predictions.
-    return predict(mach, X)
-end
-
-# =========================
-# Neural Network Models (White-box)
-# =========================
-
-"""
-    SimpleConvBuilder
-
-Minimal convolutional network builder for grayscale images such as MNIST.
-This is intended as a white-box baseline for gradient-based attacks.
-"""
-struct SimpleConvBuilder
-    filter_size::Int
-    channels1::Int
-    channels2::Int
-end
-
-"""
-    FlattenLayer(flat_dim::Int)
-
-Named flatten layer for reliable serialization.
-Reshapes 4D image tensor (H, W, C, N) to 2D matrix (flat_dim, N).
-"""
-struct FlattenLayer
-    flat_dim::Int
-end
-
-(f::FlattenLayer)(x) = reshape(x, f.flat_dim, size(x, 4))
-
-"""
-    MLJFlux.build(b::SimpleConvBuilder, rng, n_in, n_out, n_channels)
-
-Build small ConvNet for image classification.
-Assumes 28×28 input images.
-"""
-function MLJFlux.build(b::SimpleConvBuilder, rng, n_in, n_out, n_channels)
-    k, c1, c2 = b.filter_size, b.channels1, b.channels2
-    @assert isodd(k) "filter_size must be odd."
-
-    p = div(k - 1, 2)
-    init = Flux.glorot_uniform(rng)
-
-    # Calculate flattened dimension after 2 MaxPool layers
-    h, w = 28, 28
-    h, w = div(h, 2), div(w, 2)   # First MaxPool
-    h, w = div(h, 2), div(w, 2)   # Second MaxPool
-    flat_dim = h * w * c2
-
-    return Chain(
-        Conv((k, k), n_channels => c1, pad = (p, p), relu, init = init),
-        MaxPool((2, 2)),
-        Conv((k, k), c1 => c2, pad = (p, p), relu, init = init),
-        MaxPool((2, 2)),
-        FlattenLayer(flat_dim),
-        Dense(flat_dim, 128, relu, init = init),
-        Dense(128, n_out, init = init),
-        # NOTE: MLJFlux adds softmax automatically
-    )
-end
-
-"""
-    make_mnist_cnn(; rng=42, epochs=5, batch_size=64)
-
-Create MLJFlux ImageClassifier for MNIST.
-"""
-function make_mnist_cnn(; rng::Int = 42, epochs::Int = 5, batch_size::Int = 64)
-    builder = SimpleConvBuilder(3, 16, 32)
-
-    model = ImageClassifier(
-        builder = builder,
-        loss = Flux.Losses.crossentropy,
-        optimiser = Optimisers.Adam(0.001),
-        epochs = epochs,
-        batch_size = batch_size,
-        rng = rng,
-    )
-
-    return model
-end
-
-"""
-    extract_flux_model(mach)
-
-Extract underlying Flux.Chain from MLJFlux machine.
-"""
-function extract_flux_model(mach)
-    fp = fitted_params(mach)
-    return fp.chain
 end
 
 # =========================
@@ -386,7 +249,7 @@ flux_model = extract_flux_model(mach)
 
 """
 function get_or_train_cnn(name::String = "mnist_cnn"; force_retrain = false, kwargs...)
-    \ # Check cache
+    # Check cache
     if !force_retrain
         cached = load_experiment_result(name)
         if !isnothing(cached)
